Refactoring AuthService into interface and implementation
