+++ wmsdev/wmsweb/src/main/java/com/wmsweb/service/AuthService.java
@@ -1,0 +1,11 @@
+package com.wmsweb.service;
+
+import org.springframework.stereotype.Service;
+
+import wmsdata.UserBean;
+
+public interface AuthService {
+
+	boolean authenticate(UserBean bean);
+
+}
